digital-twin-workshop/requirements.txt
fastapi
uvicorn
python-dotenv
upstash-vector
groq
